feat: merge and check duplicates
